+++ src/main/java/com/minji/idusbackend/cbn/CabinetProductService.java
@@ -7,6 +7,7 @@ import com.minji.idusbackend.cbn.model.PatchCbnRes;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import java.math.BigInteger;
 import java.util.List;
 
 
@@ -15,19 +16,19 @@ public class CabinetProductService {
     @Autowired
     private CabinetProductDao cabinetProductDao;
 
-    public String addCbn(int userLoginResIdx, String cbn_name) {
+    public String addCbn(BigInteger userLoginResIdx, String cbn_name) {
         return cabinetProductDao.addCbn(userLoginResIdx, cbn_name);
     }
 
-    public List<GetMyCbnList> cbnList(int userLoginResIdx) {
+    public List<GetMyCbnList> cbnList(BigInteger userLoginResIdx) {
         return cabinetProductDao.cbnList(userLoginResIdx);
     }
 
-    public List<GetOneCbnRes> getCbn(int userLoginResIdx, int cabinet_idx) {
+    public List<GetOneCbnRes> getCbn(BigInteger userLoginResIdx, int cabinet_idx) {
         return cabinetProductDao.getCbn(userLoginResIdx, cabinet_idx);
     }
 
-    public PatchCbnRes updateCbn(int userLoginResIdx, int product_idx, String cbn_idx) throws Exception {
+    public PatchCbnRes updateCbn(BigInteger userLoginResIdx, int product_idx, String cbn_idx) throws Exception {
         try {
             return cabinetProductDao.updateCbn(userLoginResIdx, product_idx, cbn_idx);
         } catch (Exception exception) {
