+++ src/main/java/com/minji/idusbackend/member/model/UserLoginRes.java
@@ -5,17 +5,18 @@ import lombok.Setter;
 import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.userdetails.User;
 
+import java.math.BigInteger;
 import java.util.Collection;
 
 @Getter
 @Setter
 
 public class UserLoginRes extends User {
-    int idx;
+    BigInteger idx;
     String email;
     String nickname;
 
-    public UserLoginRes(int idx, String username, String password, String nickname, Collection<?
+    public UserLoginRes(BigInteger idx, String username, String password, String nickname, Collection<?
             extends GrantedAuthority> authorities) {
         super(username, password, authorities);
         this.idx = idx;
