+++ src/main/java/com/minji/idusbackend/pay/PayService.java
@@ -8,6 +8,7 @@ import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.math.BigInteger;
 import java.util.List;
 
 @Service
@@ -16,7 +17,7 @@ public class PayService {
     @Autowired
     PayDao payDao;
 
-//    public String chargePay(int userLoginRes, int money) {
+//    public String chargePay(BigInteger userLoginRes, int money) {
 //        return payDao.chargePay(userLoginRes, money);
 //    }
 
@@ -24,19 +25,19 @@ public class PayService {
 //        return payDao.showTotalPay(member_idx);
 //    }
 
-    public List<PostPayRes> showPayChargeList(int member_idx){
+    public List<PostPayRes> showPayChargeList(BigInteger member_idx){
         return payDao.showPayChargeList(member_idx);
     }
 
-    public String chargePay2(int userLoginRes, int money) {
+    public String chargePay2(BigInteger userLoginRes, int money) {
         return payDao.chargePay2(userLoginRes, money);
     }
 
-    public int showTotalPay2(int member_idx){
+    public int showTotalPay2(BigInteger member_idx){
         return payDao.showTotalPay2(member_idx);
     }
 
-    public String withdrawPay(int userLoginRes, int money) {
+    public String withdrawPay(BigInteger userLoginRes, int money) {
         return payDao.withdrawPay(userLoginRes, money);
     }
